requirements.txt
# Agentic RAG System Requirements
# 向量数据库和检索系统依赖

# 核心依赖
camel-ai>=0.1.5
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
jieba>=0.42.1

# 基础科学计算包
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 机器学习和深度学习
torch>=2.0.0
transformers>=4.21.0
huggingface-hub>=0.15.0

# 向量数据库
qdrant-client>=1.7.0

# 文本处理
tokenizers>=0.13.0
datasets>=2.0.0

# 网络和异步
aiohttp>=3.8.0
httpx>=0.24.0

# 解决依赖冲突的包
absl-py>=0.15.0
grpcio>=1.48.2
six>=1.16.0
protobuf>=3.20.0

# 可选：Jupyter支持
# jupyter>=1.0.0
# ipywidgets>=7.6.0

# 可选：可视化
# matplotlib>=3.5.0
# seaborn>=0.11.0

# 可选：进度条
tqdm>=4.64.0

# 可选：配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# Reranking (Pinecone Inference API)
pinecone>=5.0.0